requirements.txt
asgiref==3.2.10
bcrypt==3.2.0
cffi==1.14.2
cryptography==3.1
digi-xbee==1.3.0
Django==3.1
django-cors-headers==3.5.0
django-sslserver==0.22
future==0.18.2
iso8601==0.1.12
mariadb==1.0.0
mysql-connector-python==8.0.21
mysqlclient==2.0.1
paramiko==2.7.2
pkg-resources==0.0.0
protobuf==3.13.0
pycparser==2.20
PyMySQL==0.10.0
PyNaCl==1.4.0
pyserial==3.4
pytz==2020.1
PyYAML==5.3.1
RPi.GPIO==0.7.0
six==1.15.0
sqlparse==0.3.1
srp==1.0.16
sshtunnel==0.1.5
XBee==2.3.2
